refactor(directories): redesign the pages structure to support to new flow
