--- ShoppingApp/app/(tabs)/index.tsx
@@ -1,23 +1,0 @@
-import { Text, View, StyleSheet } from "react-native";
-
-export default function Index() {
-  return (
-    <View
-      style={styles.container}
-    >
-      <Text style={styles.text}>Edit app/index.tsx to edit this screen.</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#25292e',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  text: {
-    color: '#fff',
-  },
-});
