--- ShoppingApp/app/(tabs)/shopping_list.tsx
@@ -1,24 +1,0 @@
-import { Text, View, StyleSheet } from 'react-native';
-import { GestureHandlerRootView } from 'react-native-gesture-handler';
-
-import ShoppingList from "@/components/ShoppingList"
-
-export default function ShoppingListScreen() {
-  return (
-    <GestureHandlerRootView style={styles.container}>
-      <ShoppingList/>
-    </GestureHandlerRootView>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#25292e',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    color: '#fff',
-  },
-});
