+++ ShoppingApp/app/index.tsx
@@ -1,0 +1,25 @@
+import { Link } from 'expo-router';
+import { View, Text, StyleSheet } from 'react-native';
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.text}>Where are you?</Text>
+      <Link href="updating">HOME</Link>
+      <Link href="shopping">STORE</Link>
+    </View>
+  );
+}
+
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#25292e',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    text: {
+        color: '#fff',
+    },
+    });
